Move range stuff in separate module
